refactor: Rename component file to avoid confusion
